Created a base class for RTF and HTML Writer
