+++ bundles/org.eclipse.swt/Eclipse SWT Custom Widgets/common/org/eclipse/swt/custom/HTMLWriter.java
@@ -14,18 +14,9 @@ import org.eclipse.swt.*;
 import org.eclipse.swt.graphics.*;
 
 /**
- * The <code>HTMLWriter</code> class is used to write widget content as
- * HTML. The implementation creates a very-very simple HTML, as compatible
+ * The <code>HTMLWriter</code> class is used to write styled content as HTML.
+ * The implementation creates a very-very simple HTML, as compatible
  * as possible: simple styles, no JavaScript.
- *
- * <p>{@code toString()} is guaranteed to return a valid formatted string only after
- * {@code close()} has been called.</p>
- *
- * <p>Whole and partial lines and line breaks can be written. Lines will be
- * formatted using the styles queried from the {@link LineStyleListener},
- * if set, or those set directly in the widget. All styles are applied to
- * the stream like they are rendered by the widget. In addition, the
- * widget font name and size is used for the whole text.</p>
  *
  * <p>How the concepts are mapped to HTML:</p>
  * <ul>
@@ -36,30 +27,13 @@ import org.eclipse.swt.graphics.*;
  *   <li>The {@link StyleRange} properties inside the line are written the {@code <span>} elements.</li>
  * </ul>
  */
-class HTMLWriter extends TextWriter {
-	private final StyledText styledText;
+class HTMLWriter extends StyledTextWriterBase {
 
-	/**
-	 * Creates a HTML writer that writes content starting at offset "start"
-	 * in the document. {@code start} and {@code length} can be set to specify
-	 * partial lines.
-	 *
-	 * @param start start offset of content to write, 0 based from beginning of document
-	 * @param length length of content to write
-	 * @param styledText the widget to produce the HTML from
-	 */
 	public HTMLWriter(StyledText styledText, int start, int length) {
-		super(start, length);
-		this.styledText = styledText;
+		super(styledText, start, length);
 		writeHeader();
 	}
 
-	/**
-	 * Closes the writer. Once closed no more content can be written.
-	 *
-	 * <p><b>NOTE:</b> {@code toString()} does not return a valid formatted string until
-	 * {@code close()} has been called.</p>
-	 */
 	@Override
 	public void close() {
 		if (!isClosed()) {
@@ -69,24 +43,8 @@ class HTMLWriter extends TextWriter {
 		}
 	}
 
-	/**
-	 * Appends the specified segment of "string" to the output data.
-	 * Copy from {@code start} up to, but excluding, {@code end}.
-	 *
-	 * @param string string to copy a segment from. Must not contain line breaks.
-	 *  Line breaks should be written using {@link #writeLineDelimiter()}
-	 * @param start start offset of segment. 0 based.
-	 * @param end end offset of segment
-	 */
-	void writeEscaped(String string, int start, int end) {
-		String textToWrite = string.substring(start, end);
-		write(escapeText(textToWrite));
-	}
-
-	/**
-	 * Writes the opening of the two {@code div}s that wrap the content.
-	 */
-	private void writeHeader() {
+	@Override
+	void writeHeader() {
 		StringBuilder outerDivStyle = new StringBuilder();
 		StringBuilder innerDivStyle = new StringBuilder();
 
@@ -139,69 +97,6 @@ class HTMLWriter extends TextWriter {
 		}
 	}
 
-	/**
-	 * Appends the specified line text to the output data. Lines will be formatted
-	 * using the styles queried from the LineStyleListener, if set, or those set
-	 * directly in the widget.
-	 *
-	 * @param line line text to write. Must not contain line breaks
-	 *  Line breaks should be written using {@link #writeLineDelimiter(String)}
-	 * @param lineOffset offset of the line. 0 based from the start of the
-	 *  widget document. Any text occurring before the start offset or after the
-	 *  end offset specified during object creation is ignored.
-	 *
-	 * @throws SWTException {@code ERROR_IO} when the writer is closed.
-	 */
-	@Override
-	public void writeLine(String line, int lineOffset) {
-		if (isClosed()) {
-			SWT.error(SWT.ERROR_IO);
-		}
-
-		int lineIndex = styledText.content.getLineAtOffset(lineOffset);
-		int lineAlignment;
-		int lineIndent;
-		boolean lineJustify;
-		int verticalIndent;
-		int[] ranges;
-		StyleRange[] styles;
-
-		StyledTextEvent event = styledText.getLineStyleData(lineOffset, line);
-		if (event != null) {
-			verticalIndent = event.verticalIndent;
-			lineAlignment = event.alignment;
-			lineIndent = event.indent;
-			lineJustify = event.justify;
-			ranges = event.ranges;
-			styles = event.styles;
-		} else {
-			verticalIndent = styledText.renderer.getLineVerticalIndent(lineIndex);
-			lineAlignment = styledText.renderer.getLineAlignment(lineIndex, styledText.alignment);
-			lineIndent = styledText.renderer.getLineIndent(lineIndex, styledText.indent);
-			lineJustify = styledText.renderer.getLineJustify(lineIndex, styledText.justify);
-			ranges = styledText.renderer.getRanges(lineOffset, line.length());
-			styles = styledText.renderer.getStyleRanges(lineOffset, line.length(), false);
-		}
-
-		if (styles == null) {
-			styles = new StyleRange[0];
-		}
-
-		event = styledText.getLineBackgroundData(lineOffset, line);
-		Color lineBackground = (event != null && event.lineBackground != null)
-				? event.lineBackground
-				: styledText.renderer.getLineBackground(lineIndex, null);
-
-		writeStyledLine(line, lineOffset, ranges, styles, lineBackground, lineIndent, verticalIndent, lineAlignment, lineJustify);
-	}
-
-	/**
-	 * Appends the specified line delimiter to the output data.
-	 *
-	 * @param lineDelimiter line delimiter to write as RTF.
-	 *
-	 * @throws SWTException {@code ERROR_IO} when the writer is closed.
-	 */
 	@Override
 	public void writeLineDelimiter(String lineDelimiter) {
 		if (isClosed()) {
@@ -210,87 +105,7 @@ class HTMLWriter extends TextWriter {
 		write(lineDelimiter);
 	}
 
-	/**
-	 * Appends the specified line text to the output data.
-	 *
-	 * <p>Use the colors and font styles specified in {@code styles} and {@code lineBackground}.
-	 * Formatting is written to reflect the text rendering by the text widget.
-	 * Style background colors take precedence over the line background color.</p>
-	 *
-	 * @param line line text to write as RTF. Must not contain line breaks
-	 *  Line breaks should be written using writeLineDelimiter()
-	 * @param lineOffset offset of the line. 0 based from the start of the
-	 *  widget document. Any text occurring before the start offset or after the
-	 *  end offset specified during object creation is ignored.
-	 * @param styles styles to use for formatting. Must not be null.
-	 * @param lineBackground line background color to use for formatting.
-	 *  May be null.
-	 */
-	private void writeStyledLine(String line, int lineOffset, int ranges[], StyleRange[] styles,
-			Color lineBackground, int indent, int verticalIndent, int alignment, boolean justify) {
-
-		int lineLength = line.length();
-		int startOffset = getStart();
-		int writeOffset = startOffset - lineOffset;
-		if (writeOffset >= lineLength) {
-			return;
-		}
-		int lineIndex = Math.max(0, writeOffset);
-
-		String atLineEnd = writeLineStart(lineBackground, indent, verticalIndent, alignment, justify);
-
-		int endOffset = startOffset + super.getCharCount();
-		int lineEndOffset = Math.min(lineLength, endOffset - lineOffset);
-
-		for (int i = 0; i < styles.length; i++) {
-			StyleRange style = styles[i];
-			int start, end;
-			if (ranges != null) {
-				start = ranges[i << 1] - lineOffset;
-				end = start + ranges[(i << 1) + 1];
-			} else {
-				start = style.start - lineOffset;
-				end = start + style.length;
-			}
-			// skip over partial first line
-			if (end < writeOffset) {
-				continue;
-			}
-			// style starts beyond line end or write end
-			if (start >= lineEndOffset) {
-				break;
-			}
-			// write any unstyled text
-			if (lineIndex < start) {
-				// copy to start of style
-				// style starting beyond end of write range or end of line
-				// is guarded against above.
-				writeEscaped(line, lineIndex, start);
-				lineIndex = start;
-			}
-			// write styled text
-			String atSpanEnd = writeSpanStart(style);
-
-			// copy to end of style or end of write range or end of line
-			int copyEnd = Math.min(end, lineEndOffset);
-			// guard against invalid styles and let style processing continue
-			copyEnd = Math.max(copyEnd, lineIndex);
-			writeEscaped(line, lineIndex, copyEnd);
-
-			writeSpanEnd(atSpanEnd);
-
-			lineIndex = copyEnd;
-		}
-
-		// write the unstyled text at the end of the line
-		if (lineIndex < lineEndOffset) {
-			writeEscaped(line, lineIndex, lineEndOffset);
-		}
-
-		writeLineEnd(atLineEnd);
-	}
-
-	// Returns the text to append at the end of the line
+	@Override
 	String writeLineStart(Color lineBackground, int indent, int verticalIndent, int alignment, boolean justify) {
 		StringBuilder paragraphStyle = new StringBuilder();
 
@@ -312,14 +127,11 @@ class HTMLWriter extends TextWriter {
 			write("<p style='" + paragraphStyle + "'>");
 		}
 
+		// This is what will be used to close the line
 		return "</p>";
 	}
 
-	void writeLineEnd(String prepared) {
-		write(prepared);
-	}
-
-	// Returns the text to append at the end of the styled span
+	@Override
 	String writeSpanStart(StyleRange style) {
 		StringBuilder spanStyle = new StringBuilder();
 
@@ -399,6 +211,7 @@ class HTMLWriter extends TextWriter {
 			write("<span style='" + spanStyle2 + "'>");
 		}
 
+		// This is what will be used to close the span
 		StringBuilder toCloseSpan = new StringBuilder();
 		if (spanStyle2 != null) {
 			toCloseSpan.append("</span>");
@@ -409,13 +222,10 @@ class HTMLWriter extends TextWriter {
 		return toCloseSpan.toString();
 	}
 
-	void writeSpanEnd(String prepared) {
-		write(prepared);
-	}
-
 	// ==== Helper methods ====
 
-	private static String escapeText(String text) {
+	@Override
+	String escapeText(String text) {
 		return text
 				.replaceAll("&", "&amp;")
 				.replaceAll("\"", "&quot;")
