+++ bundles/org.eclipse.swt/Eclipse SWT Custom Widgets/common/org/eclipse/swt/custom/RTFWriter.java
@@ -19,9 +19,8 @@ import org.eclipse.swt.*;
 import org.eclipse.swt.graphics.*;
 
 /**
- * The {@code RTFWriter} class is used to write widget content as
- * rich text. The implementation complies with the RTF specification
- * version 1.5.
+ * The {@code RTFWriter} class is used to write styled content as rich text.
+ * The implementation complies with the RTF specification version 1.5.
  *
  * <p>{@code toString()} is guaranteed to return a valid formatted string only after
  * {@code close()} has been called.</p>
@@ -32,25 +31,14 @@ import org.eclipse.swt.graphics.*;
  * the stream like they are rendered by the widget. In addition, the
  * widget font name and size is used for the whole text.</p>
  */
-class RTFWriter extends TextWriter {
-	private final StyledText styledText;
+class RTFWriter extends StyledTextWriterBase {
 	static final int DEFAULT_FOREGROUND = 0;
 	static final int DEFAULT_BACKGROUND = 1;
 	List<Color> colorTable;
 	List<Font> fontTable;
 
-	/**
-	 * Creates a RTF writer that writes content starting at offset "start"
-	 * in the document. {@code start} and {@code length} can be set to specify
-	 * partial lines.
-	 *
-	 * @param start start offset of content to write, 0 based from beginning of document
-	 * @param length length of content to write
-	 * @param styledText the widget to produce the RTF from
-	 */
 	public RTFWriter(StyledText styledText, int start, int length) {
-		super(start, length);
-		this.styledText = styledText;
+		super(styledText, start, length);
 		colorTable = new ArrayList<>();
 		fontTable = new ArrayList<>();
 		colorTable.add(this.styledText.getForeground());
@@ -58,12 +46,6 @@ class RTFWriter extends TextWriter {
 		fontTable.add(this.styledText.getFont());
 	}
 
-	/**
-	 * Closes the writer. Once closed no more content can be written.
-	 *
-	 * <p><b>NOTE:</b> {@code toString()} does not return a valid formatted string until
-	 * {@code close()} has been called.</p>
-	 */
 	@Override
 	public void close() {
 		if (!isClosed()) {
@@ -74,22 +56,9 @@ class RTFWriter extends TextWriter {
 	}
 
 	/**
-	 * Appends the specified segment of "string" to the output data.
-	 * Copy from {@code start} up to, but excluding, {@code end}.
-	 *
-	 * @param string string to copy a segment from. Must not contain line breaks.
-	 *  Line breaks should be written using {@link #writeLineDelimiter()}
-	 * @param start start offset of segment. 0 based.
-	 * @param end end offset of segment
-	 */
-	void writeEscaped(String string, int start, int end) {
-		String textToWrite = string.substring(start, end);
-		write(escapeText(textToWrite));
-	}
-
-	/**
 	 * Writes the RTF header including font table and color table.
 	 */
+	@Override
 	void writeHeader() {
 		StringBuilder header = new StringBuilder();
 		FontData fontData = styledText.getFont().getFontData()[0];
@@ -132,69 +101,6 @@ class RTFWriter extends TextWriter {
 		write(header.toString(), 0);
 	}
 
-	/**
-	 * Appends the specified line text to the output data. Lines will be formatted
-	 * using the styles queried from the LineStyleListener, if set, or those set
-	 * directly in the widget.
-	 *
-	 * @param line line text to write. Must not contain line breaks
-	 *  Line breaks should be written using {@link #writeLineDelimiter(String)}
-	 * @param lineOffset offset of the line. 0 based from the start of the
-	 *  widget document. Any text occurring before the start offset or after the
-	 *  end offset specified during object creation is ignored.
-	 *
-	 * @throws SWTException {@code ERROR_IO} when the writer is closed.
-	 */
-	@Override
-	public void writeLine(String line, int lineOffset) {
-		if (isClosed()) {
-			SWT.error(SWT.ERROR_IO);
-		}
-
-		int lineIndex = styledText.content.getLineAtOffset(lineOffset);
-		int lineAlignment;
-		int lineIndent;
-		boolean lineJustify;
-		int verticalIndent;
-		int[] ranges;
-		StyleRange[] styles;
-
-		StyledTextEvent event = styledText.getLineStyleData(lineOffset, line);
-		if (event != null) {
-			verticalIndent = event.verticalIndent;
-			lineAlignment = event.alignment;
-			lineIndent = event.indent;
-			lineJustify = event.justify;
-			ranges = event.ranges;
-			styles = event.styles;
-		} else {
-			verticalIndent = styledText.renderer.getLineVerticalIndent(lineIndex);
-			lineAlignment = styledText.renderer.getLineAlignment(lineIndex, styledText.alignment);
-			lineIndent = styledText.renderer.getLineIndent(lineIndex, styledText.indent);
-			lineJustify = styledText.renderer.getLineJustify(lineIndex, styledText.justify);
-			ranges = styledText.renderer.getRanges(lineOffset, line.length());
-			styles = styledText.renderer.getStyleRanges(lineOffset, line.length(), false);
-		}
-
-		if (styles == null) {
-			styles = new StyleRange[0];
-		}
-
-		event = styledText.getLineBackgroundData(lineOffset, line);
-		Color lineBackground = (event != null && event.lineBackground != null)
-				? event.lineBackground
-				: styledText.renderer.getLineBackground(lineIndex, null);
-
-		writeStyledLine(line, lineOffset, ranges, styles, lineBackground, lineIndent, verticalIndent, lineAlignment, lineJustify);
-	}
-
-	/**
-	 * Appends the specified line delimiter to the output data.
-	 *
-	 * @param lineDelimiter line delimiter to write as RTF.
-	 *
-	 * @throws SWTException {@code ERROR_IO} when the writer is closed.
-	 */
 	@Override
 	public void writeLineDelimiter(String lineDelimiter) {
 		if (isClosed()) {
@@ -204,87 +110,7 @@ class RTFWriter extends TextWriter {
 		write("\\par ");
 	}
 
-	/**
-	 * Appends the specified line text to the output data.
-	 *
-	 * <p>Use the colors and font styles specified in {@code styles} and {@code lineBackground}.
-	 * Formatting is written to reflect the text rendering by the text widget.
-	 * Style background colors take precedence over the line background color.</p>
-	 *
-	 * @param line line text to write as RTF. Must not contain line breaks
-	 *  Line breaks should be written using writeLineDelimiter()
-	 * @param lineOffset offset of the line. 0 based from the start of the
-	 *  widget document. Any text occurring before the start offset or after the
-	 *  end offset specified during object creation is ignored.
-	 * @param styles styles to use for formatting. Must not be null.
-	 * @param lineBackground line background color to use for formatting.
-	 *  May be null.
-	 */
-	void writeStyledLine(String line, int lineOffset, int ranges[], StyleRange[] styles,
-			Color lineBackground, int indent, int verticalIndent, int alignment, boolean justify) {
-
-		int lineLength = line.length();
-		int startOffset = getStart();
-		int writeOffset = startOffset - lineOffset;
-		if (writeOffset >= lineLength) {
-			return;
-		}
-		int lineIndex = Math.max(0, writeOffset);
-
-		String atLineEnd = writeLineStart(lineBackground, indent, verticalIndent, alignment, justify);
-
-		int endOffset = startOffset + super.getCharCount();
-		int lineEndOffset = Math.min(lineLength, endOffset - lineOffset);
-
-		for (int i = 0; i < styles.length; i++) {
-			StyleRange style = styles[i];
-			int start, end;
-			if (ranges != null) {
-				start = ranges[i << 1] - lineOffset;
-				end = start + ranges[(i << 1) + 1];
-			} else {
-				start = style.start - lineOffset;
-				end = start + style.length;
-			}
-			// skip over partial first line
-			if (end < writeOffset) {
-				continue;
-			}
-			// style starts beyond line end or write end
-			if (start >= lineEndOffset) {
-				break;
-			}
-			// write any unstyled text
-			if (lineIndex < start) {
-				// copy to start of style
-				// style starting beyond end of write range or end of line
-				// is guarded against above.
-				writeEscaped(line, lineIndex, start);
-				lineIndex = start;
-			}
-			// write styled text
-			String atSpanEnd = writeSpanStart(style);
-
-			// copy to end of style or end of write range or end of line
-			int copyEnd = Math.min(end, lineEndOffset);
-			// guard against invalid styles and let style processing continue
-			copyEnd = Math.max(copyEnd, lineIndex);
-			writeEscaped(line, lineIndex, copyEnd);
-
-			writeSpanEnd(atSpanEnd);
-
-			lineIndex = copyEnd;
-		}
-
-		// write the unstyled text at the end of the line
-		if (lineIndex < lineEndOffset) {
-			writeEscaped(line, lineIndex, lineEndOffset);
-		}
-
-		writeLineEnd(atLineEnd);
-	}
-
-	// Returns the text to append at the end of the line
+	@Override
 	String writeLineStart(Color lineBackground, int indent, int verticalIndent, int alignment, boolean justify) {
 		write("\\fi");
 		write(indent);
@@ -305,14 +131,11 @@ class RTFWriter extends TextWriter {
 			write(" ");
 		}
 
+		// This is what will be used to close the line
 		return lineBackground == null ? "" : "}";
 	}
 
-	void writeLineEnd(String prepared) {
-		write(prepared);
-	}
-
-	// Returns the text to append at the end of the styled span
+	@Override
 	String writeSpanStart(StyleRange style) {
 		write("{\\cf");
 		write(getColorIndex(style.foreground, DEFAULT_FOREGROUND));
@@ -346,6 +169,7 @@ class RTFWriter extends TextWriter {
 		}
 		write(" ");
 
+		// This is what will be used to close the span
 		StringBuilder toCloseSpan = new StringBuilder();
 		if ((fontStyle & SWT.BOLD) != 0) {
 			toCloseSpan.append("\\b0");
@@ -363,13 +187,10 @@ class RTFWriter extends TextWriter {
 		return toCloseSpan.toString();
 	}
 
-	void writeSpanEnd(String prepared) {
-		write(prepared);
-	}
-
 	// ==== Helper methods ====
 
-	static String escapeText(String string) {
+	@Override
+	String escapeText(String string) {
 		StringBuilder result = new StringBuilder(string.length());
 		string.chars().forEach(ch -> {
 			if (ch > 0x7F) {
@@ -395,7 +216,7 @@ class RTFWriter extends TextWriter {
 	 * @return the index of the specified color in the RTF color table
 	 *  or "defaultIndex" if "color" is null.
 	 */
-	int getColorIndex(Color color, int defaultIndex) {
+	private int getColorIndex(Color color, int defaultIndex) {
 		if (color == null) return defaultIndex;
 		int index = colorTable.indexOf(color);
 		if (index == -1) {
@@ -413,7 +234,7 @@ class RTFWriter extends TextWriter {
 	 * @return the index of the specified color in the RTF color table
 	 *  or "defaultIndex" if "color" is null.
 	 */
-	int getFontIndex(Font font) {
+	private int getFontIndex(Font font) {
 		int index = fontTable.indexOf(font);
 		if (index == -1) {
 			index = fontTable.size();
